Redesigned layout with vector fonts
This new layout shows the time in the first row with a large vector
font. The second line shows distance and cadence with a smaller font,
while the third shows pace and heart rate with the same font as the
second. The headings have been removed.

The goal of this redesign is to increase the readability of the data
while running or hiking, with larger numbers and without unnecessary
heading texts.